refactor: move CDP runtime out of commands/dom (#218 finding #1)
Modules under src/commands/dom/ that execute CDP operations (not CLI glue)
move to src/runtime/dom/. This fixes the inverted dependency where
src/daemon/worker/ and src/ipc/protocol/ imported from @/commands/*.

After this change:
- Infrastructure layers (ipc, daemon, connection) have zero imports from
  @/commands/*
- CLI layer (src/commands/dom/*) imports runtime from @/runtime/dom/*
- Worker imports runtime from @/runtime/dom/* (not @/commands/*)

Files moved:
- evalHelpers.ts, formDiscovery.ts, formSubmitHelpers.ts, formTypes.ts
- reactEventHelpers.ts, keyMapping.ts, formFillHelpers/ (dir)

Pure rename with import path updates; no behavior change.
