+++ src/commands/dom/form.ts
@@ -7,6 +7,9 @@
 
 import type { Command } from 'commander';
 
+import { runCommand } from '@/commands/shared/CommandRunner.js';
+import { jsonOption } from '@/commands/shared/commonOptions.js';
+import { domFormDiscover } from '@/ipc/client.js';
 import type {
   FormCommandOptions,
   FormDiscoveryResult,
@@ -21,10 +24,7 @@ import type {
   RawButton,
   FieldState,
   FormFieldType,
-} from '@/commands/dom/formTypes.js';
-import { runCommand } from '@/commands/shared/CommandRunner.js';
-import { jsonOption } from '@/commands/shared/commonOptions.js';
-import { domFormDiscover } from '@/ipc/client.js';
+} from '@/runtime/dom/formTypes.js';
 import { QueryCacheManager } from '@/session/QueryCacheManager.js';
 import { formatFormDiscovery } from '@/ui/formatters/form.js';
 import { createLogger } from '@/ui/logging/index.js';
